Move components directory to root & fix relative path
