+++ components/layout.jsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import styles from "./layout.module.css";
-import utilStyles from "../../styles/utils.module.css";
+import utilStyles from "../styles/utils.module.css";
 import Link from "next/link";
 
 const name = "daido1976";
