+++ pages/index.jsx
@@ -1,5 +1,5 @@
 import Head from "next/head";
-import Layout, { siteTitle } from "./components/layout";
+import Layout, { siteTitle } from "../components/layout";
 import utilStyles from "../styles/utils.module.css";
 
 export default () => {
