+++ pages/posts/first-post.jsx
@@ -1,6 +1,6 @@
 import Head from "next/head";
 import Link from "next/link";
-import Layout from "../components/layout";
+import Layout from "../../components/layout";
 
 export default () => {
   return (
